requirements.txt
bottle==0.12.23
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
decorator==4.4.2
Flask==2.2.2
gobject==0.1.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
moviepy==1.0.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
proglog==0.1.10
proxy-tools==0.1.0
pycairo==1.23.0
PyGObject==3.42.2
pytube==12.1.2
pywebview==4.0.1
QtPy==2.3.0
requests==2.28.2
style==1.1.0
tqdm==4.64.1
update==0.0.1
urllib3==1.26.14
Werkzeug==2.2.2
